+++ lein-skeptic/dev-resources/project-runtime-reader/src/project_runtime_reader/core.clj
@@ -1,0 +1,4 @@
+(ns project-runtime-reader.core)
+
+(def registered-value
+  #date-time "2026-06-09T12:34:56Z")